Relocated Exceptions to a more appropriate package
