+++ Dockerfile
@@ -5,7 +5,7 @@ ARG PLAYA_RU_GITHUB_TOKEN
 
 ARG TMP_DIST=/tmp/keycloak
 
-ENV KEYCLOAK_VERSION 24.0.1
+ENV KEYCLOAK_VERSION 26.0.0
 ENV PLAYA_THEMES_VERSION 1.0.22
 ENV JSON_PATH_VERSION 2.7.0
 ENV JSON_SMART_VERSION 2.4.7
