+++ src/main/java/ru/playa/keycloak/modules/Utils.java
@@ -12,7 +12,7 @@ import java.util.Arrays;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.Optional;
-import ru.playa.keycloak.modules.exception.NotHostedDomainException;
+import ru.playa.keycloak.exception.NotHostedDomainException;
 
 
 /**
