+++ src/main/java/ru/playa/keycloak/modules/yandex/YandexIdentityProvider.java
@@ -14,8 +14,7 @@ import org.keycloak.models.KeycloakSession;
 import ru.playa.keycloak.modules.Utils;
 
 import java.io.IOException;
-import ru.playa.keycloak.modules.exception.MissingEmailException;
-
+import ru.playa.keycloak.exception.MissingEmailException;
 
 /**
  * Провайдер OAuth-авторизации через <a href="https://yandex.ru">Яндекс</a>.
